requirements.txt
flask
flask-cors
requests
python-dotenv
openai
supabase
psycopg2-binary
python-dateutil
gunicorn
PyJWT
cryptography
bcrypt
schedule
APScheduler>=3.10.0
# RAG System Dependencies
numpy>=1.24.0
tiktoken>=0.5.0